main_app/requirements.txt
flask==1.1.2
flask-security-too==4.0.0
flask-sqlalchemy==2.4.4
sqlalchemy-utils==0.36.8
sqlalchemy==1.3.23
flask-migrate==2.7.0
flask-mail==0.9.1
bcrypt==3.2.0
bleach==3.3.0
psycopg2-binary==2.8.6